+++ lizard_geodin/migrations/0014_auto__add_measurementconfiguration.py
@@ -1,0 +1,103 @@
+# encoding: utf-8
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        
+        # Adding model 'MeasurementConfiguration'
+        db.create_table('lizard_geodin_measurementconfiguration', (
+            ('id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('name', self.gf('django.db.models.fields.CharField')(max_length=50, null=True, blank=True)),
+            ('slug', self.gf('django.db.models.fields.SlugField')(max_length=50, db_index=True)),
+            ('measurement', self.gf('django.db.models.fields.related.ForeignKey')(blank=True, related_name=u'measurement_configurations', null=True, to=orm['lizard_geodin.Measurement'])),
+            ('metadata_filter', self.gf('jsonfield.fields.JSONField')(null=True, blank=True)),
+            ('flot_fields', self.gf('jsonfield.fields.JSONField')(null=True, blank=True)),
+        ))
+        db.send_create_signal('lizard_geodin', ['MeasurementConfiguration'])
+
+
+    def backwards(self, orm):
+        
+        # Deleting model 'MeasurementConfiguration'
+        db.delete_table('lizard_geodin_measurementconfiguration')
+
+
+    models = {
+        'lizard_geodin.apistartingpoint': {
+            'Meta': {'object_name': 'ApiStartingPoint'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'}),
+            'source_url': ('django.db.models.fields.URLField', [], {'max_length': '200', 'null': 'True', 'blank': 'True'})
+        },
+        'lizard_geodin.datatype': {
+            'Meta': {'object_name': 'DataType'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'})
+        },
+        'lizard_geodin.investigationtype': {
+            'Meta': {'object_name': 'InvestigationType'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'})
+        },
+        'lizard_geodin.locationtype': {
+            'Meta': {'object_name': 'LocationType'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'})
+        },
+        'lizard_geodin.measurement': {
+            'Meta': {'object_name': 'Measurement'},
+            'data_type': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'measurements'", 'null': 'True', 'to': "orm['lizard_geodin.DataType']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'investigation_type': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'measurements'", 'null': 'True', 'to': "orm['lizard_geodin.InvestigationType']"}),
+            'location_type': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'measurements'", 'null': 'True', 'to': "orm['lizard_geodin.LocationType']"}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '255', 'null': 'True', 'blank': 'True'}),
+            'project': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'measurements'", 'null': 'True', 'to': "orm['lizard_geodin.Project']"})
+        },
+        'lizard_geodin.measurementconfiguration': {
+            'Meta': {'object_name': 'MeasurementConfiguration'},
+            'flot_fields': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'measurement': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'measurement_configurations'", 'null': 'True', 'to': "orm['lizard_geodin.Measurement']"}),
+            'metadata_filter': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'})
+        },
+        'lizard_geodin.point': {
+            'Meta': {'object_name': 'Point'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'location': ('django.contrib.gis.db.models.fields.PointField', [], {'null': 'True', 'blank': 'True'}),
+            'measurement': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'points'", 'null': 'True', 'to': "orm['lizard_geodin.Measurement']"}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'}),
+            'source_url': ('django.db.models.fields.URLField', [], {'max_length': '200', 'null': 'True', 'blank': 'True'}),
+            'x': ('django.db.models.fields.FloatField', [], {'null': 'True', 'blank': 'True'}),
+            'y': ('django.db.models.fields.FloatField', [], {'null': 'True', 'blank': 'True'}),
+            'z': ('django.db.models.fields.FloatField', [], {'null': 'True', 'blank': 'True'})
+        },
+        'lizard_geodin.project': {
+            'Meta': {'ordering': "(u'-active', u'name')", 'object_name': 'Project'},
+            'active': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'api_starting_point': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "u'location_types'", 'null': 'True', 'to': "orm['lizard_geodin.ApiStartingPoint']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'metadata': ('jsonfield.fields.JSONField', [], {'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50', 'null': 'True', 'blank': 'True'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50', 'db_index': 'True'}),
+            'source_url': ('django.db.models.fields.URLField', [], {'max_length': '200', 'null': 'True', 'blank': 'True'})
+        }
+    }
+
+    complete_apps = ['lizard_geodin']