+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog of lizard-geodin
 0.1 (unreleased)
 ----------------
 
+- Added MeasurementConfiguration including comments on how to use it.
+
 - Added admin actions to reload the api startpoint and the projects.
 
 - Added models to match the json structure.